Convert the examples to Rmarkdown (#1)
* working on figure caption

* Change to markdown v2. Compilation with makefile.

Added .gitignore to exclude html files.
Created Makefile that uses rmarkdown to compile files with rmarkdown v2.
Change Rmd yaml header to also create .md file.

Note: on our system, the compilation does not work on bi-R-3.3.0 because of some issues with the cairo library. Use bi-R-3.1.3 to compile.

* merge all content from bioqc-kidney

* Finish conversion of bioqc-kidney.

Started working on bioqc-simulation

* working on biocq simulation. added cache folders to gitignore

* add 'simplify=TRUE' for all wmwTest calls, because of BioQC not accepting default args for certain data types

* Add next paragraph. Fix typo in simulation-benchmark filename

* add pairwise mixing subsection

* experimenting with figure references

* Complete content. Fix side-by-side-graphics. Remove _files directories.

Use the gridExtra package with grid.arrange to arrange figures side by side. This fixes the font-size-issue.

The _files directories are temporarily created by pandoc and should not be in git.

* continueing review

* continued proofreading.

* Add 'clean' to makefile

* Finish proofreading of the document

* Change 'make clean' s.t. README.md is not deleted

* Update 'make clean': increase verbosity

* Delete old Rnw and Pdf files

* Update README

* Remove *_files from gitignore.

Apparently figures are not embedded in markdown and therefore needed for proper displaying in github.

* Add filter to Makefile s.t. svg files are displayed on github

* Add comment to makefile and change the path to the 'accio' github user

* Remove rawgit subsitution command from Makefile.

We now serve contents via gh-pages. Therefore we dont't need output files
in this repo any more. Therefore
* removed all output files (including .md)
* removed the rawgit-hack from the makefile

* Adjust yaml header for gh-pages